Remove clippy-check action
Due to recent GitHub CI changes, native Dependabot no longer gets
special permissions, and is now considered the same as a third-party
forking a repository and creating a pull request, in which case they
don't have access to any secrets from the original repository.

Which raises issues with the `clippy-check` action, which uses the
`GITHUB_TOKEN` secret to have permissions to post comments on the
commit.